refactor(ink): split InkMode internals
